Changes migrations, switch decimals to floats
Accuracy of BigDecimals are not necessary and floats
are easier to use in operations

+++ simulator/db/migrate/20140926211523_create_states.rb
@@ -4,8 +4,8 @@ class CreateStates < ActiveRecord::Migration
       t.string :country_code
       t.string :name
       t.integer :population, default: 0
-      t.decimal :population_growth, default: 0.0
-      t.decimal :internet_penetration, default: 0.0
+      t.float :population_growth, default: 0.0
+      t.float :internet_penetration, default: 0.0
       t.timestamps
     end
   end
